🐛 fix(serial): 修复TransceiverManager类型转换警告
修复数据包接收循环中的类型转换问题：
- 将Capacity转换为int类型以避免有符号/无符号比较警告
- 确保边界检查的正确性

+++ hardware/serial/transceiver_manager.hpp
@@ -179,7 +179,7 @@ bool TransceiverManager<Capacity>::recv_packet(PacketType& packet) {
                 _recv_buf_len += recv_len;
 
                 // 遍历校验
-                for (int i = 0; (i + Capacity) <= _recv_buf_len; i++) {
+                for (int i = 0; (i + static_cast<int> (Capacity)) <= _recv_buf_len; i++) {
                     if (check_packet(_recv_buffer.data() + i, Capacity)) {
                         packet.copy_from(_recv_buffer.data() + i);
